% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagonal_matrix.R
\name{diagonal.matrix}
\alias{diagonal.matrix}
\title{Diagonal matrix generating function}
\usage{
diagonal.matrix(r)
}
\arguments{
\item{r}{The rank of a diagonal matrix}
}
\value{
matrix
}
\description{
Diagonal matrix generating function
}
\examples{
diagonal.matrix(5)
}
