% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonal_check.R
\name{is_orthogonal}
\alias{is_orthogonal}
\title{Check orthogonality}
\usage{
is_orthogonal(A)
}
\arguments{
\item{A}{a matrix to be checked}
}
\value{
logical
}
\description{
Check orthogonality
}
\examples{
A <- matrix(rnorm(3*4),3,4)
is_orthogonal(A)
}
