% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_matrix.R
\name{noise.matrix}
\alias{noise.matrix}
\title{Noise matrix generating function}
\usage{
noise.matrix(n, p)
}
\arguments{
\item{n}{The row size of the noise matrix}

\item{p}{The column size of the noise matrix}
}
\value{
A matrix
}
\description{
Noise matrix generating function
}
\examples{
noise.matrix(10,5)
}
