% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_norm.R
\name{op.norm}
\alias{op.norm}
\title{Operator norm}
\usage{
op.norm(A0, A1)
}
\arguments{
\item{A0}{A matrix of a fixed size, say n*p}

\item{A1}{Another matrix of the same size as A0, n*p}
}
\value{
numeric
}
\description{
Operator norm
}
\examples{
A0 <- matrix(rnorm(10*5),10,5)
A1 <- matrix(rnorm(10*5),10,5)
op.norm(A0,A1)
}
