% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonal_matrix.R
\name{orthogonal.matrix}
\alias{orthogonal.matrix}
\title{Orthogonal matrix generating function (semi-orthogonal)}
\usage{
orthogonal.matrix(n, r)
}
\arguments{
\item{n}{the row size of the generated orthogonal matrix}

\item{r}{the column size of the generated orthogonal matrix}
}
\value{
a matrix
}
\description{
Orthogonal matrix generating function (semi-orthogonal)
}
\examples{
orthogonal.matrix(10,5)
}
