% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthonormal_matrix.R
\name{orthonormal.matrix}
\alias{orthonormal.matrix}
\title{Generate orthonormal matrix}
\usage{
orthonormal.matrix(r)
}
\arguments{
\item{r}{a tuning parameter that represent the rank of the generated orthonormal matrix}
}
\value{
a matrix
}
\description{
Generate orthonormal matrix
}
\examples{
orthonormal.matrix(10)

}
