% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method_1.R
\name{SVD.est}
\alias{SVD.est}
\title{SVD estimation method}
\usage{
SVD.est(Y0, r)
}
\arguments{
\item{Y0}{The target matrix of size n*p}

\item{r}{A tuning parameter to process the rank-r estimation, r<min(n,p)}
}
\value{
A matrix
}
\description{
SVD estimation method
}
\examples{
Y0 <- matrix(rnorm(10*5),10,5)
r <- 2
SVD.est(Y0,r)

}
