% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method_3.R
\name{two.sample.weighted}
\alias{two.sample.weighted}
\title{two-sample-weighted transfer learning method}
\usage{
two.sample.weighted(w0, w1, Y0, Y1, r)
}
\arguments{
\item{w0}{The weight we put on the target data}

\item{w1}{The weight we put on the source data, note that w1=1-w0}

\item{Y0}{The target matrix of size n0*p}

\item{Y1}{The source matrix of size n1*p}

\item{r}{A tuning parameter to process the rank-r estimation, r<min(n0,n1,p)}
}
\value{
A matrix
}
\description{
two-sample-weighted transfer learning method
}
\examples{
Y0 <- matrix(rnorm(10*5),10,5)
Y1 <- matrix(rnorm(10*5),10,5)
r <- 2
w0 <- 0.3
w1 <- 1-w0
two.sample.weighted(w0,w1,Y0,Y1,r)
}
