% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method_2.R
\name{vertical.concatenation}
\alias{vertical.concatenation}
\title{Vertical concatenation transfer learning method}
\usage{
vertical.concatenation(Y0, Y1, r)
}
\arguments{
\item{Y0}{The target matrix of size n0*p}

\item{Y1}{The source matrix of size n1*p}

\item{r}{A tuning parameter to process the rank-r estimation, r<min(n0,n1,p)}
}
\value{
A matrix
}
\description{
Vertical concatenation transfer learning method
}
\examples{
Y0 <- matrix(rnorm(10*5),10,5)
Y1 <- matrix(rnorm(10*5),10,5)
r <- 2
vertical.concatenation(Y0,Y1,r)
}
